requirements.txt
pytest
playwright
requests
allure-pytest
pyyaml
loguru
